Typing

+++ irckaaja/dynamicmodule.py
@@ -1,4 +1,12 @@
 import importlib
+from typing import Any
+
+from irckaaja.botscript import BotScript
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from irckaaja.serverconnection import ServerConnection
 
 
 class DynamicModule:
@@ -6,9 +14,15 @@ class DynamicModule:
     This class holds Python scripts.
     """
 
-    def __init__(self, server_connection, modulename, config=None):
+    def __init__(
+        self,
+        server_connection: "ServerConnection",
+        module_name: str,
+        config: dict[str, Any] | None = None,
+    ) -> None:
         """
-        Initialises and calls load().
+        Initialises and tries to load a class in a module in the scripts folder.
+        Module should be named <ClassName>.lower().
 
         server_connection: connection to the network in which the module is
         related
@@ -17,36 +31,29 @@ class DynamicModule:
         instance: instance of classvar
         """
         self.server_connection = server_connection
-        self.module_name = modulename
-        self.module = None
-        self.classvar = None
-        self.instance = None
+        self.module_name = module_name
         self.module_config = config
-        self.load()
 
-    def reload_module(self):
-        """
-        Reloads the module, the class and overwrites the instance.
-        """
-        self.instance.kill()
-        importlib.reload(self.module)
+        self.module = __import__(
+            "irckaaja.scripts." + self.module_name.lower(),
+            None,
+            None,
+            [self.module_name],
+            0,
+        )
+
         self.classvar = getattr(self.module, self.module_name)
-        self.instance = self.classvar(
+        self.instance: BotScript = self.classvar(
             self.server_connection, self.module_config
         )
 
-    def load(self):
+    def reload_module(self) -> None:
         """
-        Tries to load a class in a module in the scripts folder.
-        Module should be named <ClassName>.lower().
+        Reloads the module, the class and overwrites the instance.
         """
-        self.module = __import__(
-            "scripts." + self.module_name.lower(),
-            globals(),
-            locals(),
-            [self.module_name],
-            -1,
-        )
+        if self.instance:
+            self.instance.kill()
+        importlib.reload(self.module)
         self.classvar = getattr(self.module, self.module_name)
         self.instance = self.classvar(
             self.server_connection, self.module_config
